Updates
-Finished Thermistor to what I think it should be

-Updated Some incorrect variable names

+++ libraries/main/Anemometer.h
@@ -22,8 +22,8 @@ class Anemometer : public DataSource
     double period;
 
     private:
-    double CurrentTime;
-    double oldTime;
+    int CurrentTime;
+    int oldTime;
 
 };
